requirements.txt
streamlit
pandas
langgraph
langchain
langchain_community
langchain_ollama
langchain_google_genai
langchain_openai
langchain_anthropic
duckdb
pyyaml
docker
openpyxl
redis
python-dotenv
matplotlib
seaborn
requests
